move to camera class
